fix(app-router): keep dev RSC client references valid
