+++ packages/vinext/src/plugins/rsc-client-shim-excludes.ts
@@ -3,6 +3,7 @@ const RSC_CLIENT_SHIM_OPTIMIZE_DEPS_EXCLUDE = Object.freeze([
   // bare source. If Vite pre-bundles these known client shims, the generated
   // client-package proxy can lose the matching export metadata in dev.
   "vinext/shims/error-boundary",
+  "vinext/shims/default-global-error",
   "vinext/shims/form",
   "vinext/shims/layout-segment-context",
   "vinext/shims/link",
